Move transport test to memory module
